reorganize and set up clippy
