+++ .rust-toolchain.toml
@@ -1,0 +1,3 @@
+[toolchain]
+channel = "1.61.0"
+components = ["rustfmt", "clippy"]
