+++ Cargo.toml
@@ -7,3 +7,6 @@ edition = "2021"
 
 [dependencies]
 bevy = "0.8.0"
+
+[dev-dependencies]
+rusty-hook = "0.11.2"
