+++ src/combat/components.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::*;
+
+#[derive(Component)]
+pub(crate) struct AllyCreature;
